Move classes from literals. Move module-specific constants from literals to their specific module.
